Conflicts:
	app/services/business_item/photo/storing_handler.rb

+++ app/services/business_item/payment_terms/storing_handler.rb
@@ -2,25 +2,31 @@ class BusinessItem::PaymentTerms::StoringHandler
   takes :repository, :params, :business_item
 
   def perform
-    to_create = to_create(params)
-    responses = to_create.map do |payment_term_id|
-      payment_terms_params = repository.attributes_with_business_item(
-        { payment_term_id: payment_term_id },
-        business_item
-      )
-      payment_term = repository.new(payment_terms_params)
-      validator = PaymentTermValidator.new(payment_terms_params)
-      storing_handler = ::StoringHandler.new(payment_term,
-                                           payment_terms_params,
-                                           repository,
-                                           validator)
-      storing_handler.perform
+    responses = to_create(params).map do |payment_term_id|
+      store_one(payment_term_id)
     end
-    Response.new(success: responses.all?(&:successful?), object: business_item)
+
+    success = responses.all?(&:successful?)
+    repository.delete_other_than(business_item, params) if success
+
+    Response.new(success: success, object: business_item)
+  end
+
+  def store_one(payment_term_id)
+    payment_terms_params = repository.attributes_with_business_item(
+      { payment_term_id: payment_term_id },
+      business_item
+    )
+    payment_term = repository.new(payment_terms_params)
+    validator = PaymentTermValidator.new(payment_terms_params)
+    storing_handler = ::StoringHandler.new(payment_term,
+                                          payment_terms_params,
+                                          repository,
+                                          validator)
+    storing_handler.perform
   end
 
   def to_create(params)
-    repository.delete_other_than(business_item, params)
     params.reject { |id| exising_ids.include?(id.to_i) }
   end
